+++ src/main/java/com/sparta/newsfeedapp/Service/UserService.java
@@ -1,21 +1,23 @@
-package com.sparta.newsfeedapp.Service;
+package com.sparta.newsfeedapp.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.newsfeedapp.dto.LoginRequestDto;
-import com.sparta.newsfeedapp.dto.LoginResponseDto;
-import com.sparta.newsfeedapp.dto.SignupRequestDto;
+import com.sparta.newsfeedapp.dto.userRequestDto.SignupRequestDto;
+import com.sparta.newsfeedapp.dto.userRequestDto.deleteRequestDto;
+import com.sparta.newsfeedapp.dto.userRequestDto.updateRequestDto;
+import com.sparta.newsfeedapp.dto.userResponseDto.ProfileResponseDto;
 import com.sparta.newsfeedapp.entity.User;
 import com.sparta.newsfeedapp.entity.UserStatusEnum;
 import com.sparta.newsfeedapp.jwt.JwtUtil;
 import com.sparta.newsfeedapp.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -38,6 +40,7 @@ public class UserService {
         String password = passwordEncoder.encode(requestDto.getPassword());
         String email = requestDto.getEmail();
         String name = requestDto.getName();
+        String bio = requestDto.getBio();
 
         // 회원 중복 확인
         Optional<User> checkUsername = userRepository.findByUserId(userId);
@@ -57,14 +60,70 @@ public class UserService {
         }
 
         // 사용자 등록
-        User user = new User(userId, password, email, name, UserStatusEnum.ACTIVE);
+        User user = new User(userId, password, email, name, bio, UserStatusEnum.ACTIVE);
         userRepository.save(user);
         log.info("회원가입 완료");
     }
 
-    public User loadUserByUserId(String userId) throws UsernameNotFoundException {
-        return userRepository.findByUserId(userId)
-                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다: " + userId));
+    @Transactional
+    public void deleteUser(Long id, deleteRequestDto requestDto) {
+        String userId =  requestDto.getUserId();
+        String userPassword = requestDto.getPassword();
+
+        User checkUser = loadUserById(id);
+        if (!Objects.equals(checkUser.getUserId(), userId)){
+            throw new IllegalArgumentException("유저아이디가 일치하지 않습니다.");
+        }
+        if (!passwordEncoder.matches(userPassword, checkUser.getPassword())){
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+        checkUser.deactivateUser();
+    }
+
+    //    - **프로필 조회 기능**
+//            - **사용자 ID, 이름, 한 줄 소개, 이메일**을 볼 수 있습니다.
+//            - **ID(사용자 ID X), 비밀번호, 생성일자, 수정일자**와 같은 데이터는 노출하지 않습니다.
+    public ProfileResponseDto getProfile(Long id) {
+        User checkUser = loadUserById(id);
+        String userId = checkUser.getUserId();
+        String name = checkUser.getName();
+        String email = checkUser.getEmail();
+        String bio = checkUser.getBio();
+        if (checkUser.getBio() == null) bio = "자기소개가 없습니다.";
+
+        return new ProfileResponseDto(userId,name,bio,email);
+    }
+
+    /*
+    * - 비밀번호 수정 조건
+    - 비밀번호 수정 시, 본인 확인을 위해 현재 비밀번호를 입력하여 올바른 경우에만 수정할 수 있습니다.
+    - 현재 비밀번호와 동일한 비밀번호로는 변경할 수 없습니다.
+    **  필수 예외처리
+    - 비밀번호 수정 시, 본인 확인을 위해 입력한 현재 비밀번호가 일치하지 않은 경우
+    - 비밀번호 형식이 올바르지 않은 경우
+    - 현재 비밀번호와 동일한 비밀번호로 수정하는 경우
+    * */
+    @Transactional
+    public void updateProfile(Long id, updateRequestDto requestDto) {
+        User checkUser = loadUserById(id);
+        if (!passwordEncoder.matches(checkUser.getPassword(), requestDto.getPassword())){
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+        if (!passwordEncoder.matches(checkUser.getPassword(), requestDto.getNewPassword())){
+            throw new IllegalArgumentException("현재 비밀번호와 동일한 비밀번호로는 변경할 수 없습니다.");
+        }
+        String newName = requestDto.getName();
+        String newEmail = requestDto.getEmail();
+        String newPassword = passwordEncoder.encode(requestDto.getNewPassword());
+        String newBio = requestDto.getBio();
+
+        checkUser.update(newName, newEmail, newPassword, newBio);
+    }
+
+    public User loadUserById(Long id) throws UsernameNotFoundException {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new UsernameNotFoundException("유저를 찾을 수 없습니다: " + id));
     }
 
     public void refreshToken(HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -77,16 +136,5 @@ public class UserService {
         refreshToken = authHeader.substring(7);
         userId = jwtUtil.getUserInfoFromToken(refreshToken).getId();
 
-//        if(userId != null){
-//            User user = this.userRepository.findByUserId(userId).get();
-//            if(jwtUtil.isTokenValid(refreshToken, userId)){
-//                 //accessToken 새로 발급
-//                String accessToken = jwtUtil.createToken(user.getUserId());
-//                 //refreshToken 새로 발급
-//
-//                LoginResponseDto loginResponseDto = new LoginResponseDto(accessToken, refreshToken);
-//                new ObjectMapper().writeValue(response.getOutputStream(), loginResponseDto);
-//            }
-//        }
     }
 }
